splitted our app features and state
